refactor(solidity): compiler fix

+++ contracts/MoneyToCurve.sol
@@ -16,7 +16,7 @@ import "./curvefi/ICurveFi_SwapY.sol";
 import "./curvefi/IYERC20.sol";
 
 contract MoneyToCurve is Initializable, Context, Ownable {
-    using SafeMath for uint256;
+    using SafeMath for uint;
     using SafeERC20 for IERC20;
 
     address public curveFi_Deposit;
@@ -61,10 +61,10 @@ contract MoneyToCurve is Initializable, Context, Ownable {
      * @notice Deposits 4 stablecoins (registered in Curve.Fi Y pool)
      * @param _amounts Array of amounts for CurveFI stablecoins in pool (denormalized to token decimals)
      */
-    function multiStepDeposit(uint256[4] memory _amounts) public {
+    function multiStepDeposit(uint[4] memory _amounts) public {
         address[4] memory stablecoins = ICurveFi_DepositY(curveFi_Deposit).underlying_coins();
 
-        for (uint256 i = 0; i < stablecoins.length; i++) {
+        for (uint i = 0; i < stablecoins.length; i++) {
             IERC20(stablecoins[i]).safeTransferFrom(_msgSender(), address(this), _amounts[i]);
             IERC20(stablecoins[i]).safeApprove(curveFi_Deposit, _amounts[i]);
         }
@@ -73,14 +73,14 @@ contract MoneyToCurve is Initializable, Context, Ownable {
         ICurveFi_DepositY(curveFi_Deposit).add_liquidity(_amounts, 0); //0 to mint all Curve has to
 
         //Step 2 - stake Curve LP tokens into Gauge and get CRV rewards
-        uint256 curveLPBalance = IERC20(curveFi_LPToken).balanceOf(address(this));
+        uint curveLPBalance = IERC20(curveFi_LPToken).balanceOf(address(this));
 
         IERC20(curveFi_LPToken).safeApprove(curveFi_LPGauge, curveLPBalance);
         ICurveFi_Gauge(curveFi_LPGauge).deposit(curveLPBalance);
 
         //Step 3 - get all the rewards (and make whatever you need with them)
         crvTokenClaim();
-        uint256 crvAmount = IERC20(curveFi_CRVToken).balanceOf(address(this));
+        uint crvAmount = IERC20(curveFi_CRVToken).balanceOf(address(this));
         IERC20(curveFi_CRVToken).safeTransfer(_msgSender(), crvAmount);
     }
 
@@ -88,20 +88,20 @@ contract MoneyToCurve is Initializable, Context, Ownable {
      * @notice Withdraws 4 stablecoins (registered in Curve.Fi Y pool)
      * @param _amounts Array of amounts for CurveFI stablecoins in pool (denormalized to token decimals)
      */
-    function multiStepWithdraw(uint256[4] memory _amounts) public {
+    function multiStepWithdraw(uint[4] memory _amounts) public {
         address[4] memory stablecoins = ICurveFi_DepositY(curveFi_Deposit).underlying_coins();
 
         //Step 1 - Calculate amount of Curve LP-tokens to unstake
-        uint256 nWithdraw;
-        uint256 i;
+        uint nWithdraw;
+        uint i;
         for (i = 0; i < stablecoins.length; i++) {
             nWithdraw = nWithdraw.add(normalize(stablecoins[i], _amounts[i]));
         }
 
-        uint256 withdrawShares = calculateShares(nWithdraw);
+        uint withdrawShares = calculateShares(nWithdraw);
 
         //Check if you can re-use unstaked LP tokens
-        uint256 notStaked = curveLPTokenUnstaked();
+        uint notStaked = curveLPTokenUnstaked();
         if (notStaked > 0) {
             withdrawShares = withdrawShares.sub(notStaked);
         }
@@ -116,8 +116,8 @@ contract MoneyToCurve is Initializable, Context, Ownable {
         //Step 4 - Send stablecoins to the requestor
         for (i = 0; i < stablecoins.length; i++) {
             IERC20 stablecoin = IERC20(stablecoins[i]);
-            uint256 balance = stablecoin.balanceOf(address(this));
-            uint256 amount = (balance <= _amounts[i]) ? balance : _amounts[i]; //Safepoint for rounding
+            uint balance = stablecoin.balanceOf(address(this));
+            uint amount = (balance <= _amounts[i]) ? balance : _amounts[i]; //Safepoint for rounding
             stablecoin.safeTransfer(_msgSender(), amount);
         }
     }
@@ -125,23 +125,23 @@ contract MoneyToCurve is Initializable, Context, Ownable {
     /**
      * @notice Get amount of CurveFi LP tokens staked in the Gauge
      */
-    function curveLPTokenStaked() public view returns (uint256) {
+    function curveLPTokenStaked() public view returns (uint) {
         return ICurveFi_Gauge(curveFi_LPGauge).balanceOf(address(this));
     }
 
     /**
      * @notice Get amount of unstaked CurveFi LP tokens (which lay on this contract)
      */
-    function curveLPTokenUnstaked() public view returns (uint256) {
+    function curveLPTokenUnstaked() public view returns (uint) {
         return IERC20(curveFi_LPToken).balanceOf(address(this));
     }
 
     /**
      * @notice Get full amount of Curve LP tokens available for this contract
      */
-    function curveLPTokenBalance() public view returns (uint256) {
-        uint256 staked = curveLPTokenStaked();
-        uint256 unstaked = curveLPTokenUnstaked();
+    function curveLPTokenBalance() public view returns (uint) {
+        uint staked = curveLPTokenStaked();
+        uint unstaked = curveLPTokenUnstaked();
         return unstaked.add(staked);
     }
 
@@ -156,9 +156,9 @@ contract MoneyToCurve is Initializable, Context, Ownable {
      * @notice Calculate shared part of this contract in LP token distriution
      * @param normalizedWithdraw amount of stablecoins to withdraw normalized to 18 decimals
      */
-    function calculateShares(uint256 normalizedWithdraw) internal view returns (uint256) {
-        uint256 nBalance = normalizedBalance();
-        uint256 poolShares = curveLPTokenBalance();
+    function calculateShares(uint normalizedWithdraw) internal view returns (uint) {
+        uint nBalance = normalizedBalance();
+        uint poolShares = curveLPTokenBalance();
 
         return poolShares.mul(normalizedWithdraw).div(nBalance);
     }
@@ -166,24 +166,24 @@ contract MoneyToCurve is Initializable, Context, Ownable {
     /**
      * @notice Balances of stablecoins available for withdraw
      */
-    function balanceOfAll() public view returns (uint256[4] memory balances) {
+    function balanceOfAll() public view returns (uint[4] memory balances) {
         address[4] memory stablecoins = ICurveFi_DepositY(curveFi_Deposit).underlying_coins();
 
-        uint256 curveLPBalance = curveLPTokenBalance();
-        uint256 curveLPTokenSupply = IERC20(curveFi_LPToken).totalSupply();
+        uint curveLPBalance = curveLPTokenBalance();
+        uint curveLPTokenSupply = IERC20(curveFi_LPToken).totalSupply();
 
         require(curveLPTokenSupply > 0, "No Curve LP tokens minted");
 
-        for (uint256 i = 0; i < stablecoins.length; i++) {
+        for (uint i = 0; i < stablecoins.length; i++) {
             //Get Y-tokens balance
-            uint256 yLPTokenBalance = ICurveFi_SwapY(curveFi_Swap).balances(int128(i));
+            uint yLPTokenBalance = ICurveFi_SwapY(curveFi_Swap).balances(int128(i));
             address yCoin = ICurveFi_SwapY(curveFi_Swap).coins(int128(i));
 
             //Calculate user's shares in y-tokens
-            uint256 yShares = yLPTokenBalance.mul(curveLPBalance).div(curveLPTokenSupply);
+            uint yShares = yLPTokenBalance.mul(curveLPBalance).div(curveLPTokenSupply);
 
             //Get y-token price for underlying coin
-            uint256 yPrice = IYERC20(yCoin).getPricePerFullShare();
+            uint yPrice = IYERC20(yCoin).getPricePerFullShare();
 
             //Re-calculate available stablecoins balance by y-tokens shares
             balances[i] = yPrice.mul(yShares).div(1e18);
@@ -193,12 +193,12 @@ contract MoneyToCurve is Initializable, Context, Ownable {
     /**
      * @notice Balances of stablecoins available for withdraw normalized to 18 decimals
      */
-    function normalizedBalance() public view returns (uint256) {
+    function normalizedBalance() public view returns (uint) {
         address[4] memory stablecoins = ICurveFi_DepositY(curveFi_Deposit).underlying_coins();
-        uint256[4] memory balances = balanceOfAll();
+        uint[4] memory balances = balanceOfAll();
 
-        uint256 summ;
-        for (uint256 i = 0; i < stablecoins.length; i++) {
+        uint summ;
+        for (uint i = 0; i < stablecoins.length; i++) {
             summ = summ.add(normalize(stablecoins[i], balances[i]));
         }
         return summ;
@@ -207,14 +207,14 @@ contract MoneyToCurve is Initializable, Context, Ownable {
     /**
      * @notice Util to normalize balance up to 18 decimals
      */
-    function normalize(address coin, uint256 amount) internal view returns (uint256) {
+    function normalize(address coin, uint amount) internal view returns (uint) {
         uint8 decimals = ERC20Detailed(coin).decimals();
         if (decimals == 18) {
             return amount;
         } else if (decimals > 18) {
-            return amount.div(uint256(10)**(decimals - 18));
+            return amount.div(uint(10)**(decimals - 18));
         } else if (decimals < 18) {
-            return amount.mul(uint256(10)**(18 - decimals));
+            return amount.mul(uint(10)**(18 - decimals));
         }
     }
 }
