+++ contracts/curvefi/ICurveFi_Minter.sol
@@ -10,7 +10,7 @@ interface ICurveFi_Minter {
 
     function mint_for(address gauge_addr, address _for) external;
 
-    function minted(address _for, address gauge_addr) external view returns (uint256);
+    function minted(address _for, address gauge_addr) external view returns (uint);
 
     function toggle_approve_mint(address minting_user) external;
