+++ contracts/curvefi/IYERC20.sol
@@ -2,20 +2,20 @@ pragma solidity ^0.5.17;
 
 contract IYERC20 {
     //ERC20 functions
-    function totalSupply() external view returns (uint256);
+    function totalSupply() external view returns (uint);
 
-    function balanceOf(address account) external view returns (uint256);
+    function balanceOf(address account) external view returns (uint);
 
-    function transfer(address recipient, uint256 amount) external returns (bool);
+    function transfer(address recipient, uint amount) external returns (bool);
 
-    function allowance(address owner, address spender) external view returns (uint256);
+    function allowance(address owner, address spender) external view returns (uint);
 
-    function approve(address spender, uint256 amount) external returns (bool);
+    function approve(address spender, uint amount) external returns (bool);
 
     function transferFrom(
         address sender,
         address recipient,
-        uint256 amount
+        uint amount
     ) external returns (bool);
 
     function name() external view returns (string memory);
@@ -25,11 +25,11 @@ contract IYERC20 {
     function decimals() external view returns (uint8);
 
     //Y-token functions
-    function deposit(uint256 amount) external;
+    function deposit(uint amount) external;
 
-    function withdraw(uint256 shares) external;
+    function withdraw(uint shares) external;
 
-    function getPricePerFullShare() external view returns (uint256);
+    function getPricePerFullShare() external view returns (uint);
 
     function token() external returns (address);
 }
